refactor(tui): split sql_completion into keywords + refs modules
