+++ httui-tui/src/sql_completion/keywords.rs
@@ -1,0 +1,211 @@
+/// ANSI-ish keyword set shared across dialects. Covers the 80% the
+/// average note will type. Sorted, deduped, all-uppercase canonical
+/// form — the popup match is case-insensitive so `select` still
+/// matches `SELECT`.
+pub(super) const ANSI_KEYWORDS: &[&str] = &[
+    "ADD",
+    "ALL",
+    "ALTER",
+    "AND",
+    "ANALYZE",
+    "AS",
+    "ASC",
+    "BEGIN",
+    "BETWEEN",
+    "BY",
+    "CASCADE",
+    "CASE",
+    "CAST",
+    "CHECK",
+    "COLUMN",
+    "COMMIT",
+    "CONSTRAINT",
+    "CREATE",
+    "CROSS",
+    "DEFAULT",
+    "DELETE",
+    "DESC",
+    "DISTINCT",
+    "DROP",
+    "ELSE",
+    "END",
+    "EXCEPT",
+    "EXISTS",
+    "EXPLAIN",
+    "FALSE",
+    "FOREIGN",
+    "FROM",
+    "FULL",
+    "GROUP",
+    "HAVING",
+    "IF",
+    "IN",
+    "INDEX",
+    "INNER",
+    "INSERT",
+    "INTERSECT",
+    "INTO",
+    "IS",
+    "JOIN",
+    "KEY",
+    "LEFT",
+    "LIKE",
+    "LIMIT",
+    "NOT",
+    "NULL",
+    "OFFSET",
+    "ON",
+    "OR",
+    "ORDER",
+    "OUTER",
+    "PRIMARY",
+    "REFERENCES",
+    "RIGHT",
+    "ROLLBACK",
+    "SELECT",
+    "SET",
+    "TABLE",
+    "THEN",
+    "TRUE",
+    "UNION",
+    "UNIQUE",
+    "UPDATE",
+    "USING",
+    "VALUES",
+    "VIEW",
+    "WHEN",
+    "WHERE",
+    "WITH",
+];
+
+/// Postgres-flavored extras — keywords + dialect-specific syntax
+/// like `RETURNING` and `ILIKE` that aren't in pure ANSI but show up
+/// constantly in real notes.
+pub(super) const POSTGRES_KEYWORDS: &[&str] = &["ILIKE", "MATERIALIZED", "RECURSIVE", "RETURNING"];
+
+/// MySQL extras — `IGNORE`, `REPLACE`, etc. Conservative list; the
+/// engine accepts anything but we keep the popup focused on what
+/// users actually type.
+pub(super) const MYSQL_KEYWORDS: &[&str] = &["IGNORE", "REPLACE", "STRAIGHT_JOIN"];
+
+/// SQLite extras — `PRAGMA` is the big one; the rest mirror common
+/// dialect-specific syntax.
+pub(super) const SQLITE_KEYWORDS: &[&str] = &["AUTOINCREMENT", "GLOB", "PRAGMA", "VACUUM"];
+
+/// Postgres function builtins. Curated, not exhaustive — covers
+/// aggregates, string manipulation, JSON, and the date/time helpers
+/// users reach for daily.
+pub(super) const POSTGRES_FUNCTIONS: &[&str] = &[
+    "ABS",
+    "AVG",
+    "CASE",
+    "COALESCE",
+    "COUNT",
+    "CURRENT_DATE",
+    "CURRENT_TIMESTAMP",
+    "DATE_PART",
+    "DATE_TRUNC",
+    "EXTRACT",
+    "GENERATE_SERIES",
+    "GREATEST",
+    "INITCAP",
+    "JSONB_BUILD_OBJECT",
+    "JSONB_EACH",
+    "JSONB_EXTRACT_PATH",
+    "LEAST",
+    "LENGTH",
+    "LOWER",
+    "MAX",
+    "MIN",
+    "NOW",
+    "NULLIF",
+    "POSITION",
+    "REGEXP_REPLACE",
+    "REPLACE",
+    "ROUND",
+    "ROW_NUMBER",
+    "STRING_AGG",
+    "SUBSTRING",
+    "SUM",
+    "TO_CHAR",
+    "TO_DATE",
+    "TO_TIMESTAMP",
+    "TRIM",
+    "UPPER",
+];
+
+/// MySQL function builtins.
+pub(super) const MYSQL_FUNCTIONS: &[&str] = &[
+    "ABS",
+    "AVG",
+    "CONCAT",
+    "CONCAT_WS",
+    "COUNT",
+    "CURDATE",
+    "CURRENT_DATE",
+    "CURRENT_TIMESTAMP",
+    "DATE_ADD",
+    "DATE_FORMAT",
+    "DATE_SUB",
+    "DAY",
+    "EXTRACT",
+    "GREATEST",
+    "GROUP_CONCAT",
+    "HOUR",
+    "IF",
+    "IFNULL",
+    "JSON_EXTRACT",
+    "JSON_OBJECT",
+    "LEAST",
+    "LENGTH",
+    "LOWER",
+    "MAX",
+    "MIN",
+    "MONTH",
+    "NOW",
+    "NULLIF",
+    "REPLACE",
+    "ROUND",
+    "ROW_NUMBER",
+    "SUBSTRING",
+    "SUM",
+    "TIMESTAMP",
+    "TRIM",
+    "UPPER",
+    "YEAR",
+];
+
+/// SQLite function builtins. Smaller list — SQLite has fewer
+/// builtins. Includes `JSON_EXTRACT` for the JSON1 extension which
+/// is commonly enabled.
+pub(super) const SQLITE_FUNCTIONS: &[&str] = &[
+    "ABS",
+    "AVG",
+    "CASE",
+    "COALESCE",
+    "COUNT",
+    "DATE",
+    "DATETIME",
+    "GROUP_CONCAT",
+    "IFNULL",
+    "JSON_EXTRACT",
+    "JULIANDAY",
+    "LENGTH",
+    "LIKELY",
+    "LOWER",
+    "MAX",
+    "MIN",
+    "NULLIF",
+    "PRINTF",
+    "RANDOM",
+    "REPLACE",
+    "ROUND",
+    "STRFTIME",
+    "SUBSTR",
+    "SUM",
+    "TIME",
+    "TRIM",
+    "TYPEOF",
+    "UNLIKELY",
+    "UPPER",
+];
